Fix: VM/Accelerator: redundant accessing to RHS operands of AND/OR instructions
When results of AND/OR instructions are obvious from values of LHS operands
by the short-circuit evaluation, accessing to RHS should be skipped,
because data of RHS operand could not be allocated/initialized in such case.

However, before this commit, BoolScalarLogicalUnit in the accelerator
had always accessed to data of RHS operands of AND/OR instructions.
The above incorrect behaviour has been fixed by this commit.