Fix: VM/Processor: Conversion from STRING to INT64 by CAST instruction
Before this commit, when a STRING value to be converted
to a INT64 value contains a decimal point,
CAST instruction had ignored the content after the decimal point.
However, an exponent part (e.g.: "E4" of "1.23E4") might be
contained after the decimal point, and it should not be ignored.

After this commit, in the above case,
such STRING value will converted to a FLOAT64 value at first,
and then converted to a INT64 value,
so its exponent part will not be ignored.