Remove: Memory: DataContainer.setData(T data) method
The method DataContainer.setData(T data) to set "data" field
does not initialize "lengths" and "offset" fields
which are deeply related with "data" field.
They should be set together for preventing
miss initialization and keeping consistency.

In addition, the signature of setData(T data) method is ambiguous
for recognizing whether the caller of the method want to handle
the data container as a scalar or as a vector(array),
from the code of caller-side.

Therefore, we have made setData(T data) method private
on DataContainer class, and have removed it from ADCI1 interface
which is implemented by DataContainer class.

After this commit, the caller should call
setData(T data, int offset) for setting a scalar value, or
setData(T data, int[] lengths) for setting a vector value,
instead of calling setData(T data) method.