Fix: Compiler: redundant and incorrect re-allocations of registers
Before this commit, generated code for every binary operators
always had contained an ALLOC instruction for allocating a register
at which the evaluated value of the operator will be stored,
when the "ASSEMBLY_VALUE" attribute of the AST node is a register.
However, the register can be allocated/initialized at other place.

For example: the register for storing the evaluated value of
an compound-assignment operator "+="
is the same with it of the LHS operand of the operator,
and when the LHS operand is an element of a vector(array),
that register will be initialized by an ELEM instruction
generated before the generating code of the "+=" operator.

In the above case, if an ALLOC instruction will be generated
after the ELEM instruction, at runtime the data of the register will be
re-allocated and the reference to the element of the vector
from the register will be disposed, so the evaluated value
of "+=" operator will not be wrote-back to the element of the vector.
The above behaviour is incorrect.

After this commit, to fix the above incorrect behaviour,
an AST node representing an operator has "NEW_REGISTER" attribute,
if the register for storing its evaluated value requires to be newly allocated.
In addition, the ALLOC instruction for allocating the register will be generated
for an AST node only when the AST node has "NEW_REGISTER" attribute.