Add: VM: new operation code "ENDFUN"
This new instruction will be generated at the tail of code of a function,
for detecting that the execution flow of the function which should return a value
has reached to the end of the function without returning a value
(= without executing RET instructions).

When this instruction is executed, an error to notify the above cause will occur,
and the execution of code will end.