Change/Fix: AstNode: how to set "depth" and "blockDepth" fields
Before this commit, in AstNode class,
"depth" and "blockDepth" fields had been set implicitly in "addChildNode(AstNode)"
method, by computing from values of them of the parent node.

However, constructing of the AST might be incompleted yet
at the compiling stage using the above method,
and the depth of each node depends on the whole completed shape of the AST,
so the above specification might be cause of incorrect state.

After this commit, values of them will NOT be set in addChildNode(AstNode) method,
and they will be set by calling updateDepths() method of the instance at the ROOT
of the AST explicitly, from outside of the AstNode class.
It should be called after when construction of the AST has completed.
In the compiler, it will be called at the end of the parsing stage in Parser class.