Change: VM: rename operation code "AND"/"OR" to "ANDM"/"ORM"
These instructions does not access to "inputB"
when the result can be determined from only the value of "inputA",
to simplify the generated code of short-circuit evaluations of
"&&" and "||" operators.

However, this behaviour will probably be cause of bottlenecks of
performances of vector operations in future,
and then variations of these instructions with ommitting the above behaviour
will probably be added, for using in cases which don't require to perform
short-circuit evaluations.
Those new instructions should have more simple names than current
(with short-circuit evaluation behaviour) instructions,
because their behaviour are more simple than behavour of current instructions.

Therefore, we renamed current instructions from "AND"/"OR" to "ANDM"/"ORM",
for using names of "AND"/"OR" for new (simple versions of) instructions in future.