Add: VM: new operation code "ALLOCT"
This instruction is a variation of the ALLOC instruction, to only declare
data type and array rank/lengths for readability and optimizability of code,
without allocating the actual memory.

For an example of use,
this instruction will be used at the point at which data will be popped from the stack.
Data passed through the stack depends on the processing flow of the code on runtime,
so it is not easy to grasp/infere type of popped data (containing the array-rank)
without executing code is not easy.

Therefore, in such cases, declaring type/rank/lengths of data by this instruction
is helpful for readers of code and optimizers.
(the compiler knows the type of data to be popped from the stack,
so it can generate this instruction at the above point.)