Add: Compiler: support function-call-by-reference
When "&" is described before the name of an argument,
e.g.: "void fun(int &x) {...}",
the data of the argument will be passed by reference.

The generated code can already run on the VM/Processor.

However, it can't run on VM/Accelerator yet,
because an optimization-related problem occurs,
so it requires to update VM/Accelerator.