Fix: Compiler: miss-detection of ends of blocks
Before this commit, when the Semantic Analyzer was traversing nodes in the AST,
ends of blocks had been detected by finding changing points of "block-depth"
(which is hierarchical depth about nests of blocks).
However, "block-depth" does not change at the end of blocks in some cases
(e.g.: "... } { ..." ), so they could not detected.

In addition, in the preorder Depth-First-Traversal (preorder DFT),
a parent node will be visited only before traversing their children nodes,
so it will not be visited when the traversing route goes out of it.
Therefore, in some cases (e.g.: "... } }"),
multiple closing points of blocks can be contained in 1-step of
the traversing route of the preorder DFT.
The old implementation had not considered the above fact.

By considering how we can handle well above cases and more complicated cases,
we felt that the above old algorithm (containing modified version) is not appropriate
for detecting ends of blocks, so we determined to quit to use it.

As an alternative way, we extended the "getPreorderDftNextNode" method of AstNode class
for enabling to detect/notify closed node on the route to the next node.

Above problems have been fixed by this commit, by using the above new method.