Add(large): make language-spec customizable without rebuilding
Originally, this project "Vnano" started as a subproject of
the programming language VCSSL,
for developing compact & modular engine for embedded use in applications,
so the language-spec of Vnano (= VCSSL nano) is a subset of VCSSL's spec.

In near future,
the development of next generation VCSSL Engine will (probably) start.
To avoid maintaining almost duplicate source code in multiple projects,
we want to use a customized (extended) version of Vnano Engine
as a part of the next generation VCSSL Engine.

However, Before this commit, settings of Vnano Engine about language-spec
(data type names, keywords, operator precedences, and so on)
had defined in setting classes as "static final" fields.
The above means that, for customize language-spec settings,
Vnano Engine had require modifications of its source code,
and also require to be rebuilt for reflecting their changings.

On the above old way, it should be troublesome to apply updating
of the original Vnano Engine to the customized Vnano Engine.

Therefore, by this commit,
we have removed "static final" from fields in setting-classes, and
have made the reference to them dynamic in all classes in Compiler/VM.

After this commit, we can create instances of setting-classes
and modify their fields (or inherit them and override their methods),
and we can pass them to the constructor of VnanoEngine class.
On the above new way, we can customize the language-spec
without modifying original source code in the Vnano Engine,
and it does not require to rebuild the Vnano Engine.