Add: PluginLoader/LibraryLoader: improve error handlings
Before this commit: an exception was thrown
just when the loading of an library/plugin was failed,
so remained libraries/plugins in the loading list were
not loaded even if they are loadable.

After this commit: If any library/plugin isn't loadable,
an exception will be thrown AFTER when loadings of all
loadable plugins/libraries will have been completed
(so only unloadable plug-ins will be skipped).