Add: interconnect: permission-based security system
By this commit, the permission-based security system
(like as it of the VCSSL Engine) has been implemented.

Plug-ins can request permissions through requestPermission(string)
method of org.vcssl.connect.EngineConnectorInterface1 interface,
which instance will be passed to the parameter of
initializeForConnection(...) method of plug-ins.

The engine allows/denies permission-requests, by referencing values
in the "permission map", which is a Map storing names and values
of permission items. The names and the values are defined in:
org.vcssl.connect.ConnectorPermissionName & ConnectorPermissionValue.

The default value of the permission map is "ASK" for all items,
so engine asks the user whether allows the permission-request or not.

If you want to use a customized permission map,
pass it to put(...) method of VnanoScriptEngine instances
with the special key "___VNANO_PERMISSION_MAP",
or pass it to setPermissionMap(...) method of VnanoEngine instances.