Change: some example script code in README: use "print" instead of "output"
Because how contents passed to \"output\" function will be displayed
depends on host-applications, and some GUI applications (e.g. RINPn)
might not display multi-line texts as output.

"output" function is for submiting a result value to the host-application,
and "print" function is for printing texts in general purpose.

Therefore, "print" is more suitable for some example script code in README.