Fix: interconnect: initializations of XNCI plug-ins
Before this commit, when a XNCI1 plug-in was connected,
preInitializationForConnection() was called twice
(should be called only once),
and postInitializationForConnection() was not called
(should be called).

This incorrect behaviour was caused by the modification
of the commit 6116fb277eb7bd27a3a16dc24de51a2931218707.
It has been fixed by this commit.