Refactor: some example code
On the Vnano, "double" is the alias of "float",
supported only for the purpose of making it easy for users
to port code snipets written in some other languages.

Breafly, "float" is standard,
and using "double" is a special case.

For the above reason, the use of "double"
in official example code is not appropreate.
For users, it may cause unnecessary confusion about
which is better, "float" or "double" ?

Therefore,
we have replaced "double" in all example code to "float".