Add: Accelerator: execute ELEM instruction without bypassing it to the Processor
Before this commit,
ELEM instructions of the VM was very heavy
so it had been bottlenecks of processing speeds of
scripts using arrays.
The above had been caused by that:
ELEM instructions had not been executed by units in
the Accelerator, so it had been bypassed to the Processor.

To eliminate the above bottleneck, by this commit,
we have implemented some units for
executing ELEM instructions, in the Accelerator.

As the result,
on "BubbleSort.vnano" under the condition of N=3000,
the execution frequency of ELEM instructions
(measured by --perf option on the command line mode)
has improved from over 80% to about 35%,
and the IPS (instructions per second) ratio ot the VM
has improved to about 3~ times faster.

We think that we can improve the performance of
array access more, but it requires to add some
new instruction(s) and support it by the compiler.
This commit also contains some preparations for the above
(some new units in the Accelerator are not used yet).