Add: Accelerator: support NEG instructions
Before this commit,
NEG instructions had not been processed in the Accelerator,
so it had been bypassed to the Processor.

After this commit, NEG istructions will be processed in the
Accelerator directly, whthout bypassing it to the Processor.

As the result of the above, the frequency of NEG instructions
in periodically sampled instructions (by using --perf option)
has improved from over 25% to less than 3% for the same script.