Copy: Accelerator: from AcceleratorSchedulingUnit to AcceleratorOptimizationUnit
"AcceleratorSchedulingUnit" (scheduler) was originally
the unit to determine which instructions should be executed by
which operation unit (we refer this procedure as "scheduling").

However, as the result of pursuing more efficient scheduling
for improving the processing speed, most part of code of
the current scheduler is filled by code for "optimization".

Therefore, to make clear which unit actually do what,
we want to separate code of the current scheduler into 2 classes:
"AcceleratorSchedulingUnit" (current scheduler class)
and "AcceleratorOptimizationUnit" (new class).

As the first step, we have copied whole code of
the current scheduler to the new class by this commit.

After the next commit, we will remove duplicate code
between above classes, and then, for using them independently,
we will modify front-ends and back-ends of them.