Add: Accelerator: detect unnecessary MOV instructions and remove them
Sometimes intermediate code (VRIL code) generated by
the compiler of the Vnano contains MOV instructions
of which output values (stored to registers) are unused.

After this commit, such unnecessary MOV instructions
will be removed in the Accelerator before executing code.

For example,
the evaluation value of a post increment operator
in a expression (e.g.: "a = i++" ) should be the value of
the operand variable before when it is incremented,
so the compiler generates the following code
for a post increment operator:

  Firstly, stores the value of the operand variable (e.g.: i)
  to a register by using a MOV instruction, and then
  increments the operand variable by using a ADD instruction.

On the other hand,
probably, most of increment operators will be used
for incrementing counters in "for" statements, like as:
"for (int i=0; i<n; i++) {...}".

In such case, the evaluation value of "i++"
will not be used, so we can omit the MOV instruction
storing the value of the variable "i" to a register.