Add: Accelerator: execute a couple of comparison & branch instructions
In most cases, loop-condition expressions of
"for" loops are very simple comparison operations,
e.g.: "i<n" of "for (int i=0; i<n; i++) {...}".

In the compiled code of the above "for" loop,
the comparison operation "i<n" is compiled
to one comparison instruction of the VM.
After the comparison instruction,
one branch instruction is put,
for breaking from the loop when the value of
"i<n" is "false".

( Strictly speaking, in the compiled code,
  a LABEL instruction is put between
  the above comparison & branch instructions,
  but it will be removed by optimizations. )

Hence, couples of comparison & branch instructions
are executed very frequently when the VM execute
compiled code of scripts having "for" loops.

After this commit, for reducing overheads of "for" loops,
the VM detects couples of comparison & branch instructions,
and converts each couple in them to one fused instruction,
and execute it in one cycle (step) of the VM.