Add: assignment operations from an array to a scalar
In the VCSSL, the assignment operations from an array
to a scalar variable are syntactically valid.
For example:

  scalar = array ;

In the above, when the size of the "array" is 1,
the value of array[0] will be assigned to
the "scalar" variable, otherwise an error will occur.

Syntax of the Vnano is a subset of the VCSSL,
so we have implemented the above feature by this commit.