Fix: re-initialization of variables by default values
In the Vnano, variables are initialized automatically
by default values at declaration statements.

For example:

  for (int i=0; i<n; i++) {
      int v;
      output(v);
      v = 123;
  }

In the above code,
the value of v will be initialized by 0 repeatedly
at the first statement in the for-loop,
so output value will always be 0.

However, before this commit,
if the accelerator option was enabled,
a variable was initialized only at the first time
its declaration statement was evaluated.
So, the output value of the above example code was 0
only at the first cycle of the for-loop,
and was 123 after the next cycle.

The above incorrect behaviour has been fixed by this commit.