Fix: DataContainer: implementations of scalar data accessors
Scalar data accessors of DataContainer had not considered
that the container may reference to other container.

Caused by the above, when data referencing to other data
was/were passed to an external function which accesses to args
through scalar data accessors, unexpected errors occurred.

For example:

  void internal_function(float &arg) {
      external_function(arg);
  }
  internal_function(1.23);

When the above code was executed,
if external_function(float) accesses to data of
the "arg" through Float64ScalarDataAccessorInterface,
errors with the message "no data is stored" occurred.

The above incorrect behaviour has been fixed by this commit.