Add: in Accelerator: new execution units for executing ALLOC/ALLOCR instructions
Before this commit,
ALLOC/ALLOCR instructions had been bypassed to the Processor,
so its overhead costs had been larger than other instructions
which are executed in the Accelerator directly.

Those overhead costs had suppressed performance
when the optimization level is 0 or 1.

After this commit, ALLOC/ALLOCR instructions will be executed
by units in the Accelerator directly,
so performance has improved especially when
the optimization level is 0 or 1.