Refactor: change fields of some config classes to "static final"
Fields of following config classes had orinigally been
"static final", but they had changed to instance fields
by commit 9d51ff1085b6b2e852295e651dcc4dd1b71bb8c6 etc.

* ScriptWord
* AssemblyWord
* DataTypeName
* LiteralSyntax
* IdentifierSyntax

As explained by the above commit comment,
we had planed to divert this Vnano Engine as a core module of
the next generation VCSSL Engine, so we removed "static final"
from fields of above config classes to make them "swappable".

(Note: language specification of Vnano is a subset of VCSSL.)

However, it was found that the above plan increases
complexity of the Vnano Engine more than we thought,
and it does not decreases the complexity of VCSSL Engine
as much as we thought.

Hence, we have judged that there is unacceptable disadvantage
in the above plan, even we can avoid duplication of some code.

Therefore we determined to abandon the above plan,
and changed fields of config classes to "static final" again
to make code of Vnano Engine simpler.

(So we will implement the next generation VCSSL Engine
 independently of Vnano Engine.)