Change: VnanoEngine: types of exceptions thrown by some methods
The following methods of the VnanoEngine class
had been declared to throw VnanoException (checked exception):

* terminateScript()
* resetTerminator()
* getPermissionMap()
* getPerformanceMap()

And the following method had been declared to throw no exception:

* getOptionMap()

The type of exception thrown by the above methods
has been changed by this commit,
to VnanoFatalException, which is an unchecked exception.

Because:

We had added some methods for confirming
whether above methods is available.
So apps should check it before calling the above methods,
for example:

    if (engine.isTerminatorEnabled()) {
        engine.terminateScript();
    }

The app-side can reliably prevent the occurrence of the exception
by checking the availability as the above.

Hence, we have changed the type of the exception from
"checked" VnanoException to "unchecked" VnanoFatalException.

We want to restrict the use (throwing) of checked exceptions only for
cases that app-side can not prevent reliably the occurance of it,
even if the implementation of the app is perfect.
(For example, an syntax error in a script inputted by an user.)