Fix: validations of XFCI1 plg-ins: about types of return values
When the automatic data conversion feature is disabled,
the getReturnUnconvertedClass() method will not be called.
However, the validation process always checked it,
and raised an error if the value of the above method is null.

It is unnatural restriction that a plug-in must return
non-null value even when it provides a void-type function.

We have fixed the above behaviour by this commit.