Refactor: VM/ExecutionUnit: remove /* and */ enclosing reord(...) method
It will be used when our VM supports "REORD" instruction
in future, but the instruction has not been supported yet,
so this method had been commented-out.

But the long commented-out code is ambiguous
wether it is a old code so we can remove it,
or it will be used in future so we should not remove it.

Hence, we have removed /* and */ from the "reord()" method,
and put the short comment at the top of the method.