Fix: error detection for = operation between arrays
Assignment operations (=) between arrays must raise errors
when the rank (the number of dimensions) of the left operand
is different from the rank of right operand.
For example:

  float a[2][3];
  float b[6];
  a = b;

However, before this commit,
the compiler in the script engine had not raised any errors
for the above operations.
It has been fixed by this commit.