Change: VnanoEngine: rename activateEngine() to activate(), etc.
We added the following methods recently, and it was officially released yesterday (Ver.1.1.0).

* VnanoEngine.activateEngine()
* VnanoEngine.deactivateEngine()

When we determined these names, we considered that we should prioritize
clearity (what they does) than simplicity, for naming of methods of this class.
Because VnanoEngine class is the front-end class of this library,
so it has many methods such as executeScript(...), connectPlugin(...), and so on.

However, the word "Engine" at the end of the above method names are too redundant,
because they are methods of the engine itself.

We feel that the above names are strange now,
so we renamed them as follows, and re-release this library today (Ver.1.1.2).

* VnanoEngine.activate()
* VnanoEngine.deactivate()