Fix: DataConverter: fall-through
When we added a new data type in future and then
we forget to implement "case" for it,
this fall-through may be cause of a bug that
this method returns a data of which number of dimensions is incorrect.

But the above does not occur on the current code, so this is a "stealth bug".