Fix: build scripts: compile plug-in interfaces explicitly
The list of plug-in interfaces in
  src/org/vcssl/connect/
directory had not been specified explicitly in build.sh and build.bat,
so some of them referenced from the engine had been compiled,
but the other had not been compiled.

(The uncompiled interfaces had been loaded dynamically
 from an other place: plugin/org/vcssl/connect/*,
 in which all compiled interfaces exists for implementing plug-ins.)

On the other hand, when we use build.xml and Ant or other building tool,
all of the interfaces in src/org/vcssl/connect/ directory are compiled.

Hence, if we once run build.xml (or use other building environment),
and then use build.bat/sh after some time has passed,
some of the interfaces are recompiled and the others are remains old.
This can be cause of very confusing compatibility problem.

To fix the above, we have updated build.sh and build.bat
to compile all plug-in interfaces explicitly.